Refactor: Move filter to different package
https://github.com/metasfresh/metasfresh/issues/6682
